PLT-3482: Added shrink
